Modularize 2020 plugin
