+++ plugin-2020/src/game/mod.rs
@@ -1,0 +1,24 @@
+//! The game structures for the "Hive" game.
+//! Source: Partially translated from https://github.com/CAU-Kiel-Tech-Inf/socha/blob/8399e73673971427624a73ef42a1b023c69268ec/plugin/src/shared/sc/plugin2020/util/GameRuleLogic.kt
+
+mod board;
+mod constants;
+mod field;
+mod game_move;
+mod game_state;
+mod piece_type;
+mod piece;
+mod player_color;
+mod player;
+mod positioned_field;
+
+pub use board::*;
+pub use constants::*;
+pub use field::*;
+pub use game_move::*;
+pub use game_state::*;
+pub use piece_type::*;
+pub use piece::*;
+pub use player_color::*;
+pub use player::*;
+pub use positioned_field::*;
