+++ plugin-2020/src/game/player.rs
@@ -1,0 +1,19 @@
+use socha_client_base::{util::SCResult, xml_node::FromXmlNode, xml_node::XmlNode};
+
+use super::PlayerColor;
+
+/// Metadata about a player.
+#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+pub struct Player {
+    pub color: PlayerColor,
+    pub display_name: String
+}
+
+impl FromXmlNode for Player {
+    fn from_node(node: &XmlNode) -> SCResult<Self> {
+        Ok(Self {
+            color: node.attribute("color")?.parse()?,
+            display_name: node.attribute("displayName")?.to_owned()
+        })
+    }
+}
